requirements.txt
flask==2.0.3
Werkzeug==2.0.3
pillow==9.0.0
numpy==1.21.6
scikit-learn==0.24.2
joblib==1.1.0
gunicorn==20.1.0